Interpretation of Bayes factors

Quantify the strength of the evidence provided by the data to a model/hypothesis according a Bayes factor interpretation scale suggested by Kass and Raftery (1995) .

bfactor_interpret_kr(bf)

Arguments

bf
A numeric vector of non-negative values.

Value

Returns a character vector with the same length as bf.

Details

Bayes factors are a summary of the evidence provided by the data to a model/hypothesis. Because it can be useful to consider twice the natural logarithm of the Bayes factor, which is in the same scale as the familiar deviance and likelihood ratio test statistics, Kass and Raftery (1995) suggested the following Bayes factor interpretation scale:

2*log(Bayes factor)	Bayes factor	Evidence

[-Inf, 0[	[0, 1[	Negative

[0, 2[	[1, 3[	Weak

[2, 6[	[3, 20[	Positive

[6, 10[	[20, 150[	Strong

[10, +Inf[	[150, +Inf[	Very strong

bfactor_interpret_kr takes Bayes factors (bf) as input and returns the strength of the evidence in favor of the model/hypothesis in the numerator of the Bayes factors (usually the null hypothesis) according to the aforementioned table.

When comparing results with those from standard likelihood ratio tests, it is convenient to put the null hypothesis in the denominator of the Bayes factor so that bfactor_interpret_kr returns the strength of the evidence against the null hypothesis. If bf was obtained with the null hypothesis on the numerator, one can use bfactor_interpret_kr(1/bf) to obtain the strength of the evidence against the null hypothesis.

References

Kass RE, Raftery AE (1995). “Bayes factors.” Journal of the American Statistical Association, 90(430), 773--795.

See also

Examples

# Interpretation of one Bayes factor
bfactor_interpret_kr(1.5)

#> [1] "Weak"

# Interpretation of many Bayes factors
bfactor_interpret_kr(c(0.1, 1.2, 3.5, 13.9, 150))

#> [1] "Negative"    "Weak"        "Positive"    "Positive"    "Very Strong"

# Application: chi-squared goodness-of-fit test.
# Strength of the evidence provided by the lower
# bound on the Bayes factor in favor of the null hypothesis:
x <- matrix(c(12, 15, 14, 15), ncol = 2)
bfactor_interpret_kr(bcal(chisq.test(x)[["p.value"]]))

#> [1] "Weak"