Check if an object is a vector of valid probability values

check_prob checks if an object is a numeric vector of valid probability values. This can be useful to validate inputs, intermediate calculations or outputs in user-defined functions.

check_prob(p, allow_nas = TRUE)

Arguments

p
An arbitrary object.

allow_nas
Logical value indicating either TRUE or FALSE. If FALSE the execution is stopped (via stop) in case there are NA or NaN values in p.

Value

check_prob does not return any output. There are three possible scenarios:

The call is silent if p is a numeric vector of valid probability values without NA or NaN values.

An informative warning message is given if p is a numeric vector of valid probability values with some NA or NaN values and allow_nas is set to TRUE.

An informative error message is thrown and the execution is stopped if:

p is not a numeric vector of valid probability values.

p is a numeric vector of valid probability values with some NA or NaN values and allow_nas is set to FALSE.

Details

check_prob conducts a series of tests to check if p is a numeric vector of valid probability values. Namely, check_prob checks if:

p is NULL or empty.

p is a numeric (atomic) vector.

p has NA or NaN values.

The values of p are in the [0, 1] interval.

See also

Examples

# Calls that pass silently:
check_prob(0)
check_prob(0.5)
check_prob(1)
check_prob(c(0.1, 0.2, 0.3, 0.4, 0.5))

# Call that throws an informative warning message:
if (FALSE) check_prob(c(0.1, 0.2, NA, 0.4, 0.5))
if (FALSE) check_prob(c(0.1, 0.2, NA, 0.4, 0.5), allow_nas = TRUE)

# Calls that throw informative error messages:
if (FALSE) check_prob(NULL)
if (FALSE) check_prob(TRUE)
if (FALSE) check_prob("0.5")
if (FALSE) check_prob(factor(0.5))
if (FALSE) check_prob(matrix(0.5))
if (FALSE) check_prob(list(0.5))
if (FALSE) check_prob(NA)
if (FALSE) check_prob(NaN)
if (FALSE) check_prob(1.1)
if (FALSE) check_prob(-0.5)
if (FALSE) check_prob(c(-0.9, 0, 0.1, 0.2, 0.3, 0.4, 0.5))
if (FALSE) check_prob(c(0.1, 0.2, NA, 0.4, 0.5), allow_nas = FALSE)